requirements.txt
aniso8601==8.1.0
attrs==20.3.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==7.1.2
Flask==1.1.2
flask-marshmallow==0.14.0
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.4
google-api-core==2.11.0
google-api-python-client==2.77.0
google-auth==2.16.0
google-auth-httplib2==0.1.0    
google-auth-oauthlib==1.0.0    
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.1
oauthlib==3.2.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.17.3
pytz==2020.4
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.15.0
SQLAlchemy==1.3.20
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==0.16.1
zipp==3.4.0
